refactor: move tokenlist util functions to new folder
