+++ scripts/merge-tokenlists.ts
@@ -2,15 +2,16 @@ import fs from "fs";
 
 import { List, Network } from "../src/types";
 import { TokenList } from "@uniswap/token-lists";
+import { mergeTokenLists } from "../src/tokenlists/merge";
 
 async function run() {
   try {
-    await mergeTokenLists(
+    await mergeTokenListsByPath(
       `generated/${Network.Homestead}.${List.Listed}.tokenlist.json`,
       `generated/${Network.Kovan}.${List.Listed}.tokenlist.json`,
       `generated/${List.Listed}.tokenlist.json`
     );
-    await mergeTokenLists(
+    await mergeTokenListsByPath(
       `generated/${Network.Homestead}.${List.Vetted}.tokenlist.json`,
       `generated/${Network.Kovan}.${List.Vetted}.tokenlist.json`,
       `generated/${List.Vetted}.tokenlist.json`
@@ -21,7 +22,7 @@ async function run() {
   }
 }
 
-async function mergeTokenLists(
+async function mergeTokenListsByPath(
   mergerPath: string,
   mergedPath: string,
   outputPath: string
@@ -32,10 +33,7 @@ async function mergeTokenLists(
   const mergedFile = await fs.readFileSync(mergedPath);
   const merged: TokenList = JSON.parse(mergedFile.toString());
 
-  const output: TokenList = {
-    ...merger,
-    tokens: [...merger.tokens, ...merged.tokens],
-  };
+  const output: TokenList = mergeTokenLists(merger, merged);
 
   await fs.writeFileSync(outputPath, JSON.stringify(output, null, 4));
 }
